tms-24 feat: Сделал пример для UsersStorage
